// src/utils/types/global.ts
export type TProductSize = 'S' | 'M' | 'L' | 'XL' | '2XL' | '3XL'

export type TProductColor = {
  title: string
  value: string
}

export type TPrintAreaInfo = {
  // productImageId: string
  id: number
  area: {
    printX?: number // pixel position x
    printY?: number // pixel position y
    printW?: number // pixel width
    printH?: number // pixel height
    widthRealPx: number // real width of product image (for exporting)
    heightRealPx: number // real height of product image (for exporting)
  }
  surfaceType: TSurfaceType
  imageUrl: string
}

export type TProductMockupInfo = {
  variantId: number
  surfaceId: number
  mockupUrl: string
}

export type TBaseProduct = {
  id: number
  url: string
  name: string
  description: string
  variants: TClientProductVariant[] // variants
  detailImages: string[]
  inNewLine: boolean
  printAreaList: TPrintAreaInfo[] // surfaces
  mockups: TProductMockupInfo[]
}

export type TProductCategory =
  | 'shirt'
  | 'hat'
  | 'cup'
  | 'keychain'
  | 'phonecase'
  | 'figurine'
  | 'totebag'

export type TClientProductVariant = {
  id: number
  name: string
  size: TProductSize
  color: TProductColor
  priceAmountOneSide: number
  priceAmountBothSide: number
  currency: string
  priceAfterDiscount?: number
  stock: number
  category?: TProductCategory
}
// ...
export type TElementVisualBaseState = {
  position: {
    x: number
    y: number
  }
  scale: number
  angle: number
  zindex: number
}

export type TTextVisualState = Omit<TElementVisualBaseState, 'scale'> & {
  id: string
  fontSize: number
  textColor: string
  content: string
  fontFamily: string
  fontWeight: number
}

export type TStickerVisualState = TElementVisualBaseState & {
  id: string
  path: string
}

export type TPrintedImageVisualState = TElementVisualBaseState & {
  id: string
  url: string
}

export type TElementsVisualState = Partial<{
  stickers: TStickerVisualState[]
  printedImages: TPrintedImageVisualState[]
  texts: TTextVisualState[]
}>

export type TMockupDataId = string

export type TMockupData = {
  id: string
  elementsVisualState: TElementsVisualState
  imageData: TMockupImageData
  surfaceInfo: {
    id: number
    type: TSurfaceType
  }
  preSentImageLink?: string
}

export type TProductInCart = TProductCartInfo & {
  mockupDataList: TMockupData[]
}

export type TSurfaceType = 'front' | 'back' | 'both'
// ...
export type TImgMimeType = 'image/png' | 'image/jpeg' | 'image/webp'